requirements.txt
# Core Django
Django>=5.1,<6.0
djangorestframework>=3.15
gunicorn>=23.0

# Database
psycopg2-binary>=2.9

# Redis & Task Queue
redis>=5.0
django-redis>=5.4
django-rq>=2.10

# Authentication & Security
PyJWT>=2.9

# HTTP & Requests
requests>=2.32
urllib3>=2.2
certifi>=2024.1

# Payment Processors
stripe>=11.0
squareup>=39.0

# Forms
django-simple-captcha>=0.6
Pillow>=11.0

# Select2 widget
django-select2>=8.2

# Utilities
python-dateutil>=2.9
pytz>=2024.1
